# docs/swagger.yaml
basePath: /api
definitions:
  models.AdminUpdateClientDTO:
    properties:
      billing_software:
        type: string
      citizen_card_expiry:
        type: string
      citizen_card_number:
        type: string
      date_of_birth:
        type: string
      email:
        type: string
      fiscal_address:
        description: Morada fiscal
        type: string
      fiscal_city:
        type: string
      fiscal_county:
        type: string
      fiscal_district:
        type: string
      fiscal_postal_code:
        type: string
      fixed_phone:
        type: string
      marital_status:
        type: string
      name:
        description: Dados pessoais que admin pode editar
        type: string
      nif:
        type: string
      official_email:
        description: Configurações
        type: string
      phone:
        type: string
      preferred_contact_hours:
        type: string
      preferred_format:
        type: string
      status:
        type: string
    type: object
  models.AdminUpdateCompanyDTO:
    properties:
      accounting_regime:
        description: Regimes
        type: string
      address:
        description: Morada da empresa
        type: string
      annual_revenue:
        type: number
      bank_name:
        type: string
      bic:
        type: string
      business_activity:
        description: Operacionais
        type: string
      cae:
        type: string
      city:
        type: string
      company_name:
        type: string
      corporate_object:
        type: string
      county:
        type: string
      district:
        type: string
      estimated_revenue:
        type: number
      founding_date:
        type: string
      has_stock:
        type: boolean
      iban:
        type: string
      legal_form:
        type: string
      main_clients:
        type: string
      main_suppliers:
        type: string
      monthly_invoices:
        type: integer
      nipc:
        type: string
      number_employees:
        type: integer
      postal_code:
        type: string
      report_frequency:
        type: string
      share_capital:
        description: Dados financeiros
        type: number
      trade_name:
        type: string
      vat_regime:
        type: string
    type: object
  models.ApprovalRequestDTO:
    properties:
      request_id:
        example: 1
        type: integer
      review_notes:
        example: Documentação em ordem
        type: string
      status:
        enum:
        - approved
        - rejected
        example: approved
        type: string
    required:
    - request_id
    - status
    type: object
...
  models.SuccessResponse:
    properties:
      data: {}
      message:
        example: Operação realizada com sucesso
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  models.UpdateUserStatusDTO:
    properties:
      notes:
        example: Motivo da alteração
        type: string
      status:
        enum:
        - approved
        - rejected
        - blocked
        example: approved
        type: string
    required:
    - status
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Sistema de gestão contabilística com aprovação de clientes
  title: RV Contabilidade API
  version: "1.0"
paths:
  /admin/approve-request:
    post:
      consumes:
      - application/json
      description: Aprova ou rejeita uma solicitação de registo (apenas contabilistas/admin)
      parameters:
      - description: Dados de aprovação
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ApprovalRequestDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
      security:
      - BearerAuth: []
      summary: Aprovar/rejeitar solicitação
      tags:
      - admin
  /admin/clients:
    get:
      consumes:
      - application/json
      description: Lista todos os clientes com status aprovado (apenas contabilista/admin)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
      security:
      - BearerAuth: []
      summary: Listar clientes aprovados
      tags:
      - admin
  /admin/clients/{id}:
    delete:
      consumes:
      - application/json
      description: Elimina um cliente e a sua empresa (apenas contabilista/admin)
      parameters:
      - description: ID do cliente
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
      security:
      - BearerAuth: []
      summary: Eliminar cliente
      tags:
      - admin
    put:
      consumes:
      - application/json
      description: Atualiza dados pessoais de um cliente (apenas contabilista/admin)
      parameters:
      - description: ID do cliente
        in: path
        name: id
        required: true
        type: integer
      - description: Dados para atualizar
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.AdminUpdateClientDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
      security:
      - BearerAuth: []
      summary: Atualizar dados de cliente
      tags:
      - admin
  /admin/clients/{id}/company:
    put:
      consumes:
      - application/json
      description: Atualiza dados da empresa de um cliente (apenas contabilista/admin)
      parameters:
      - description: ID do cliente
        in: path
        name: id
        required: true
        type: integer
      - description: Dados da empresa para atualizar
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.AdminUpdateCompanyDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
      security:
      - BearerAuth: []
      summary: Atualizar empresa do cliente
      tags:
      - admin
  /admin/pending-requests:
    get:
      consumes:
      - application/json
      description: Lista todas as solicitações de registo pendentes (apenas contabilistas/admin)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
      security:
      - BearerAuth: []
      summary: Listar solicitações pendentes
      tags:
      - admin
  /admin/requests:
    get:
      consumes:
      - application/json
      description: Lista todos os pedidos de registo pendentes para aprovação
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
      security:
      - BearerAuth: []
      summary: Listar pedidos de registo pendentes
      tags:
      - admin
  /admin/requests/{id}:
    get:
      consumes:
      - application/json
      description: Obtém detalhes completos de um pedido de registo específico
      parameters:
      - description: ID do pedido
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
      security:
      - BearerAuth: []
      summary: Detalhes de um pedido de registo
      tags:
      - admin
  /admin/users:
    get:
      consumes:
      - application/json
      description: Lista todos os utilizadores do sistema (apenas admin/contabilista)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
      security:
      - BearerAuth: []
      summary: Listar todos os utilizadores
      tags:
      - admin
  /admin/users/{id}:
    get:
      consumes:
      - application/json
      description: Obtém detalhes completos de um utilizador
      parameters:
      - description: ID do utilizador
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
      security:
      - BearerAuth: []
      summary: Detalhes de um utilizador
      tags:
      - admin
  /admin/users/{id}/status:
    put:
      consumes:
      - application/json
      description: Bloqueia/desbloqueia utilizador (apenas admin)
      parameters:
      - description: ID do utilizador
        in: path
        name: id
        required: true
        type: integer
      - description: Novo status
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateUserStatusDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
      security:
      - BearerAuth: []
      summary: Alterar status de utilizador
      tags:
      - admin
  /admin/users/count:
    get:
      consumes:
      - application/json
      description: Conta quantos utilizadores existem por status e role
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
      security:
      - BearerAuth: []
      summary: Contar utilizadores
      tags:
      - admin
  /admin/users/simple:
    get:
      consumes:
      - application/json
      description: Lista utilizadores sem relacionamentos para debug
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
      security:
      - BearerAuth: []
      summary: Listar utilizadores (simples)
      tags:
      - admin
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"